+++ torch/quantization/ns/utils.py
@@ -6,7 +6,7 @@ import torch.nn as nn
 toq = torch.ops.quantized
 from torch.fx import GraphModule
 from torch.fx.graph import Node
-from torch.quantization.fx.quantize import is_activation_post_process
+from torch.quantization.quantize import is_activation_post_process
 
 from .ns_types import NSNodeTargetType
